move definition to cc file
